requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
crispy-tailwind==1.0.3
Django==5.2.1
django-ajax-datatable==4.5.0
django-crispy-forms==2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
idna==3.10
numpy==2.2.5
openpyxl==3.1.5
pandas==2.2.3
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
